Fix several issues with shader on macOS
When rendering in mesh mode, use the input assembler to unpack
4-component unsigned char UNORM into half4 for the color input. Metal
doesn't allow treating the TexCoord1 register as a uint and then
shifting and masking its bits.

Make sure to do color multiplications in linear space, even if the
eventual output is in gamma space.

Metal also needs the vertex offset to be calculated in procedural
rendering mode, just like DX11 and Xbone.